Refaktorer mapping for oppdaterTilbakekrevingOppgaverApi Refaktorer til samme mønster som de andre oppdater[...]OppgaverApi
